+++ wgsl_to_wgpu/src/structs.rs
@@ -337,6 +337,7 @@ mod tests {
                 derive_encase_host_shareable: true,
                 derive_serde: false,
                 matrix_vector_types: MatrixVectorTypes::Rust,
+                rustfmt: true,
             },
         );
         let actual = quote!(#(#structs)*);
@@ -412,6 +413,7 @@ mod tests {
                 derive_encase_host_shareable: true,
                 derive_serde: true,
                 matrix_vector_types: MatrixVectorTypes::Rust,
+                rustfmt: true,
             },
         );
         let actual = quote!(#(#structs)*);
@@ -458,6 +460,7 @@ mod tests {
                 derive_encase_host_shareable: false,
                 derive_serde: false,
                 matrix_vector_types: MatrixVectorTypes::Rust,
+                rustfmt: true,
             },
         );
         let actual = quote!(#(#structs)*);
@@ -503,6 +506,7 @@ mod tests {
                 derive_encase_host_shareable: false,
                 derive_serde: false,
                 matrix_vector_types: MatrixVectorTypes::Rust,
+                rustfmt: true,
             },
         );
         let actual = quote!(#(#structs)*);
@@ -536,6 +540,7 @@ mod tests {
                 derive_encase_host_shareable: false,
                 derive_serde: false,
                 matrix_vector_types: MatrixVectorTypes::Rust,
+                rustfmt: true,
             },
         );
         let actual = quote!(#(#structs)*);